Made basic features work over TRAMP
Made type-at-point, goto definition, and repl work over TRAMP.

intero is installed, and process launched on remote machine. Other features untested at present.

Squash of the following:

Factored out call-process as separate function. Trampified process launcher.

call-process replaced by intero-call-process, which at the moment is just a wrapper for process-file (essentially a tramp aware version of call-process). start-process was replaced by start-process-file which is also tramp aware.

Made intero-make-temp-file work with tramp

intero-make-temp-file now makes temporary files on remote machines if the current buffer contains a file from a remote machine. Functions which use the temporary file don't yet cope with the remote path.

Wrote a localization function for paths

Wrote a function to localize a path to a remote machine (or return the path if not remote). This can be used to by intero to load files on a remote box, but keep the remote part in the temp filename path.

Adjust users of temporary file to relativise for host

The hostname of the temporary file is removed when it used on the remote box, and the host is added when a remote file is referenced from the local machine.

Factored out adding remote to paths

Factored out a small pattern of getting the current buffer's host and pasting in on the front of a path if it is remote, in such a way that emacs will know to feed it to TRAMP.

Got the REPL to work over TRAMP

The REPL buffer needs to know about the host for the project, but it doesn't have a file backing it, so it can't deduce the host from that. I introduced a new buffer-local variable for the host of a buffer, and rewrote the buffer host discovery functions to set this. The REPL buffer has it's buffer-local variable explicitly set, by copying the value from the buffer from which it was created.

The REPL buffer needs this information so it can cd to the correct remote path, and create temporary files in the correct place.

Small tidy

Remove a redundant function. Small simplifications + fixes.

Fixing REPL details

intero-repl-mode cleared buffer local variables, so set the buffer-host variable after intero-repl-mode is called. Wrapped the path sent to cd instead.