Made flycheck work over TRAMP
The background process was finding errors, but the flycheck backend was not identifying the buffer to display errors in. This was because it was comparing the absolute path from the buffer, and the local path from the background process. Fixed this, and provided a general utility function for comparing paths (since paths over TRAMP may differ despite representing the same file).

Did a couple of small improvements to the TRAMP compatibility code also (docstrings, factoring out TRAMP path identification).

Got flycheck to work over tramp

Flycheck was finding the errors, but because it didn't recognize the temporary file from emacs' perspective and from intero's perspective as being the same, it didn't display the errors.

Improved factoring

Made a function to compare paths, rather than relying on standardized paths being used as strings.

Tidying

Refactored path comparison function, so the "representation" function could be removed completely. Improved a couple of docstrings.